+++ src/lib/components.ts
@@ -26,13 +26,13 @@ export { default as MenuItem } from "./components/MenuItem.svelte";
 export { default as Modal } from "./components/Modal.svelte";
 export { default as Nav } from "./components/Nav.svelte";
 export { default as PageHeader } from "./components/PageHeader.svelte";
-export { default as PageSection } from "./components/PageSection.svelte";
 export { default as Popover } from "./components/Popover.svelte";
 export { default as ProgressBar } from "./components/ProgressBar.svelte";
 export { default as ProgressSteps } from "./components/ProgressSteps.svelte";
 export { default as QRCode } from "./components/QRCode.svelte";
 export { default as QRCodeReader } from "./components/QRCodeReader.svelte";
 export { default as QRCodeReaderModal } from "./components/QRCodeReaderModal.svelte";
+export { default as Section } from "./components/Section.svelte";
 export { default as Segment } from "./components/Segment.svelte";
 export { default as SegmentButton } from "./components/SegmentButton.svelte";
 export { default as SkeletonText } from "./components/SkeletonText.svelte";
